iteration of pivot
