+++ main.d.ts
@@ -98,6 +98,7 @@ export * from './dist/lib/widgets/menuItemComponent';
 export * from './dist/lib/widgets/component';
 export * from './dist/lib/widgets/menuList';
 export * from './dist/lib/widgets/popupService';
+export * from './dist/lib/widgets/componentAnnotations';
 
 export * from './dist/lib/cellNavigationService';
 export * from './dist/lib/clientExports';
