+++ src/ts/context/context.ts
@@ -283,54 +283,6 @@ function autowiredFunc(name: string, optional: boolean, classPrototype: any, met
     });
 }
 
-export function QuerySelector(selector?: string): Function {
-    return querySelectorFunc.bind(this, selector);
-}
-
-function querySelectorFunc(selector: string, classPrototype: any, methodOrAttributeName: string, index: number) {
-
-    if (selector===null) {
-        console.error('ag-Grid: QuerySelector selector should not be null');
-        return;
-    }
-    if (typeof index === 'number') {
-        console.error('ag-Grid: QuerySelector should be on an attribute');
-        return;
-    }
-
-    // it's an attribute on the class
-    var props = getOrCreateProps(classPrototype);
-    if (!props.querySelectors) {
-        props.querySelectors = [];
-    }
-    props.querySelectors.push({
-        attributeName: methodOrAttributeName,
-        querySelector: selector
-    });
-}
-
-export function EventListener(eventName?: string): Function {
-    return eventListenerFunc.bind(this, eventName);
-}
-
-function eventListenerFunc(eventName: string, target: Object, methodName: string, descriptor: TypedPropertyDescriptor<any>) {
-
-    if (eventName===null) {
-        console.error('ag-Grid: EventListener eventName should not be null');
-        return;
-    }
-
-    // it's an attribute on the class
-    var props = getOrCreateProps(target);
-    if (!props.eventListenerMethods) {
-        props.eventListenerMethods = [];
-    }
-    props.eventListenerMethods.push({
-        methodName: methodName,
-        eventName: eventName
-    });
-}
-
 export function Qualifier(name: string): Function {
     return (classPrototype: any, methodOrAttributeName: string, index: number) => {
